// frontend/web/src/services/API/proserve-wb-packaging-api/models/Recipe.ts
export interface Recipe {
    /**
     * Recipe ID
     * @type {string}
     * @memberof Recipe
     */
    recipeId: string;
    /**
     * Recipe Description
     * @type {string}
     * @memberof Recipe
     */
    recipeDescription: string;
    /**
     * Recipe Name
     * @type {string}
     * @memberof Recipe
     */
    recipeName: string;
    /**
     * Recipe Platform
     * @type {string}
     * @memberof Recipe
     */
    recipePlatform: string;
    /**
     * Operating system architecture of the recipe
     * @type {string}
     * @memberof Recipe
     */
    recipeArchitecture: string;
    /**
     * Recipe OS Version
     * @type {string}
     * @memberof Recipe
     */
    recipeOsVersion: string;
    /**
     * Status of the Recipe (CREATED/ARCHIVED)
     * @type {string}
     * @memberof Recipe
     */
    status: string;
    /**
     * Create Date
     * @type {string}
     * @memberof Recipe
     */
    createDate: string;
    /**
     * Created By
     * @type {string}
     * @memberof Recipe
     */
    createdBy: string;
    /**
     * Last Update Date
     * @type {string}
     * @memberof Recipe
     */
    lastUpdateDate: string;
    /**
     * Last Update By
     * @type {string}
     * @memberof Recipe
     */
    lastUpdatedBy: string;
}